backend/requirements.txt
annotated-types==0.7.0
anyio==4.9.0
asgiref==3.8.1
beautifulsoup4==4.13.3
cached-property==2.0.1
certifi==2025.1.31
chardet==3.0.4
charset-normalizer==3.4.1
chevron==0.14.0
click==8.1.8
deep-translator==1.11.4
distro==1.9.0
Django==5.1.7
django-cors-headers==4.7.0
django-extensions==4.1
djangorestframework==3.15.2
djangorestframework_simplejwt==5.5.0
exceptiongroup==1.2.2
frozendict==2.4.6
genanki==0.13.1
gTTS==2.5.4
gunicorn==23.0.0
h11==0.14.0
h2==3.2.0
hpack==3.0.0
hstspreload==2025.1.1
httpcore==1.0.7
httpx==0.28.1
hyperframe==5.2.0
idna==3.7
iniconfig==2.1.0
jiter==0.9.0
openai==1.68.2
packaging==25.0
pluggy==1.5.0
psycopg2-binary==2.9.10
pydantic==2.10.6
pydantic_core==2.27.2
PyJWT==2.9.0
pytest==8.3.5
pytest-django==4.8.0
python-dotenv==1.0.1
pyttsx3==2.98
PyYAML==6.0.2
requests==2.32.3
rfc3986==1.5.0
sniffio==1.3.1
soupsieve==2.6
sqlparse==0.5.3
tqdm==4.67.1
typing_extensions==4.12.2
urllib3==2.3.0
whitenoise==6.9.0
google-cloud-translate==3.20.3
